refactor(ios): split GameCatalog and GameImporter from GameLibrary
